re-order when things happen
